backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
chardet==5.2.0

# SharePoint integration
Office365-REST-Python-Client==2.5.3

# OpenAI / Azure OpenAI
openai==1.12.0

# PowerPoint generation for diagrams
python-pptx==0.6.23

# HTML/CSS parsing for HTML→PPTX conversion (skywork.ai approach)
beautifulsoup4==4.12.3
lxml==5.1.0

# Environment and utilities
python-dotenv==1.0.0
pydantic==2.5.3

